# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Machina API",
    "version": "1.0.0",
    "description": "REST API for Linux hypervisor hosts: QEMU/KVM domains via libvirt, networks, storage, host networking, backups, metrics, and optional KubeVirt bundle/apply workflows when configured."
  },
  "servers": [
    { "url": "/api/v1", "description": "Default API base" }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "machina_session",
        "description": "Cookie-based session authentication (`machina_session=`)"
      }
    },
    "schemas": {
...
      "VirtImageBuildRequest": {
        "type": "object",
        "required": ["os", "output"],
        "description": "Build a disk with virt-builder on the daemon (virt-image-build). Requires [libvirt] virt_builder_allowed. Output parent must be under libvirt pool targets or default image dirs; file must not exist.",
        "properties": {
          "os": { "type": "string", "description": "virt-builder OS template id" },
          "output": { "type": "string", "description": "Absolute path for the new image" },
          "size": { "type": "string", "default": "20G" },
          "format": { "type": "string", "default": "qcow2" },
          "hostname": { "type": "string", "default": "virtbuilder-guest.local" },
          "install": { "type": "string", "description": "Comma-separated package names for repeated --install" },
          "run_command": { "type": "array", "items": { "type": "string" } },
          "copy_in": { "type": "array", "items": { "type": "string" } },
          "firstboot_script": { "type": "string", "nullable": true, "description": "Host path to script copied into guest firstboot" },
          "root_password_file": { "type": "string", "nullable": true },
          "root_password_inline": { "type": "string", "nullable": true },
          "ssh_pubkey_file": { "type": "string", "nullable": true },
          "ssh_pubkey_inline": { "type": "string", "nullable": true },
          "update": { "type": "boolean", "default": false },
          "selinux_relabel": { "type": "boolean", "default": false },
          "extra_virt_builder_args": { "type": "array", "items": { "type": "string" } },
          "timeout_secs": { "type": "integer", "format": "int64", "default": 0, "description": "Kill virt-builder after this many seconds (0 = unlimited; daemon may set from [libvirt] virt_image_build_timeout_secs)" }
        }
      },
...
      "NetworkInfo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "uuid": { "type": "string" },
          "active": { "type": "boolean" },
          "autostart": { "type": "boolean" },
          "bridge": { "type": "string" }
        }
      },
...
      "JobKind": {
        "type": "string",
        "enum": ["virt_image_build", "vm_create", "packer_golden_build"]
      },
      "PackerGoldenBuildRequest": {
        "type": "object",
        "required": ["guest"],
        "properties": {
          "guest": {
            "type": "string",
            "description": "Script profile id for build-linux-image.sh (e.g. ubuntu2404, fedora43)",
            "example": "ubuntu2404"
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "enum": ["running", "completed", "failed"]
      },
      "JobSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "$ref": "#/components/schemas/JobKind" },
          "title": { "type": "string" },
          "status": { "$ref": "#/components/schemas/JobStatus" },
          "vm_name": { "type": "string", "nullable": true },
          "target_path": { "type": "string", "nullable": true },
          "error": { "type": "string", "nullable": true },
          "created_unix": { "type": "integer", "format": "int64" },
          "updated_unix": { "type": "integer", "format": "int64" }
        }
      },
      "JobDetail": {
        "type": "object",
        "allOf": [
          { "$ref": "#/components/schemas/JobSummary" },
          {
            "type": "object",
            "properties": {
              "logs": { "type": "array", "items": { "type": "string" } }
            }
          }
        ]
      }
    }
  },
  "paths": {
...
    "/jobs/virt-image-build": {
      "post": {
        "summary": "Start virt-image-build asynchronously (returns job id; use GET /jobs/{id} or stream)",
        "tags": ["Jobs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VirtImageBuildRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "status": { "type": "string", "example": "started" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/jobs/packer-golden-build": {
      "post": {
        "summary": "Start Golden Forge (local Packer/QEMU golden qcow2) asynchronously",
        "description": "Runs `bash /usr/local/share/machina/packer/build-linux-image.sh <guest> work` under `/var/lib/machina/packer-builds/<job-id>`. Requires the script installed (e.g. via install.sh). Returns job id; use GET /jobs/{id} or GET /jobs/{id}/stream for logs.",
        "tags": ["Jobs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PackerGoldenBuildRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "status": { "type": "string", "example": "started" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Job detail with captured log lines",
        "tags": ["Jobs"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Job detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDetail" }
              }
            }
          },
          "404": { "description": "Unknown job id" }
        }
      }
    },
    "/jobs/{id}/stream": {
      "get": {
        "summary": "SSE: log chunks (data events), then complete or error",
        "tags": ["Jobs"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
...
    "/networks": {
      "get": {
        "summary": "List all virtual networks",
        "tags": ["Networks"],
        "responses": { "200": { "description": "Network list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/NetworkInfo" } } } } } }
      },
      "post": {
        "summary": "Create a virtual network",
        "tags": ["Networks"],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "bridge_name": { "type": "string" }, "subnet": { "type": "string" }, "dhcp_start": { "type": "string" }, "dhcp_end": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Network created" } }
      }
    },
    "/networks/{name}": {
      "delete": {
        "summary": "Delete a virtual network",
        "tags": ["Networks"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Network deleted" } }
      }
    },
    "/networks/{name}/start": {
      "post": {
        "summary": "Start a virtual network",
        "tags": ["Networks"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Network started" } }
      }
    },
    "/networks/{name}/stop": {
      "post": {
        "summary": "Stop a virtual network",
        "tags": ["Networks"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Network stopped" } }
      }
    },
    "/networks/{name}/xml": {
      "get": {
        "summary": "Get network XML configuration",
        "tags": ["Networks"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Network XML" } }
      }
    },
    "/networks/{name}/autostart/{enabled}": {
      "post": {
        "summary": "Set network autostart",
        "tags": ["Networks"],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "enabled", "in": "path", "required": true, "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "Autostart updated" } }
      }
    },
...
  }
}